Moved Commands to Commands class & Removed unused using statements
[ # ] Moved all Commands to Commands class
[ - ] Removed all unused using statements
[ - ] Removed commands/
[ - ] Removed Commands.amino-help.cs
[ - ] Removed Commands.CommunityIdCommand.cs
[ - ] Removed Commands.CreatedTime.cs
[ - ] Removed Commands.DeviceCommand.cs
[ - ] Removed Commands.ExtraDeviceCommand.cs
[ - ] Removed Commands.ObjectIdCommand.cs
[ - ] Removed Commands.ProfileInfoCommand.cs
[ - ] Completely removed Commands.ProxyCheckerCommand.cs
[ - ] Removed Commands.PublicURLCommand.cs
[ - ] Removed Commands.VerifyCommand.cs
