+++ AminoBot/Events/InteractionHandler.cs
@@ -1,13 +1,5 @@
-﻿using Amino;
-using Discord;
-using Discord.Interactions;
+﻿using Discord.Interactions;
 using Discord.WebSocket;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Reflection.Metadata;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace AminoBot.Events
 {
